chore: add props and animation for timeline, shift all content const to utils and update content
