+++ components/TimelineSection.vue
@@ -1,8 +1,8 @@
 <template>
   <section id="timeline">
-    <h2>Timeline</h2>
+    <h2 class="timeline-title">My Career Journey</h2>
     <ul>
-      <li v-for="(item, index) in timeline">
+      <li v-for="(item, index) in timeline" class="timeline-item">
         <div v-if="index !== 0" class="line"></div>
         <div class="timeline-card">
           <div>
@@ -24,80 +24,40 @@
 </template>
 <script setup>
 import { CalendarIcon, LinkIcon } from "@heroicons/vue/24/outline";
-const timeline = [
-  {
-    date: "July 2023",
-    title: "Started @ JagaApp",
-    description: "Joined the JagaApp team",
-    // icon: "",
-    link: {
-      text: "Visit JagaApp",
-      url: "https://jagaapp.com/",
+import gsap from "gsap";
+import { ScrollTrigger } from "gsap/ScrollTrigger";
+gsap.registerPlugin(ScrollTrigger);
+
+const props = defineProps({
+  timeline: {
+    type: Array,
+    default: () => [],
+  },
+});
+
+onMounted(() => {
+  let titleTl = gsap.timeline({
+    scrollTrigger: {
+      trigger: ".timeline-title",
+      start: "-60px center",
+      end: "top center",
+      // markers: true,
     },
-  },
-  {
-    date: "May 2022",
-    title: "Started @ Limitless Technologies",
-    description: "Joined the Limitless Technologies (LTG) team",
-    // icon: "",
-    link: {
-      text: "Visit LTG",
-      url: "https://limitless.my/",
-    },
-  },
-  {
-    date: "October 2021",
-    title: "First Developer Role",
-    description:
-      "Joined The Hacker Collective (THC) Team as Frontend Web Developer",
-    // icon: "",
-    link: {
-      text: "Visit THC",
-      url: "https://hackercollective.co/",
-    },
-  },
-  {
-    date: "April 2021",
-    title: "Self Taught Journey Begins",
-    description:
-      "Started my journey in self learning programming (web development)",
-    // icon: "",
-  },
-  {
-    date: "August 2019",
-    title: "First Mechanical Engineer Role",
-    description:
-      "Began my job as a Mechanical Engineer in M&E Consultancy at Durianê Professional",
-    // icon: "",
-    link: {
-      text: "Visit Durianê Professional",
-      url: "https://www.duriane.com/",
-    },
-  },
-  {
-    date: "May 2019",
-    title: "Graduated University",
-    description:
-      "Completed my Bachelor's Degree in Mechanical Engineering at Multimedia University (MMU)",
-    // icon: "",
-  },
-  {
-    date: "March 2018",
-    title: "Internship",
-    description:
-      "Began an internship as assistent engineer at Protech Builders",
-    // icon: "",
-  },
-  {
-    date: "June 2015",
-    title: "Start University",
-    description:
-      "Began my study in Mechanical Engineering at Multimedia University (MMU)",
-    // icon: "",
-    link: {
-      text: "Visit MMU",
-      url: "https://www.mmu.edu.my/",
-    },
-  },
-];
+  });
+  titleTl
+    .addLabel("start")
+    .from(".timeline-title", { opacity: 0, translateX: "-100px" })
+    .addLabel("entering")
+    .to(".timeline-title", { opacity: 1, translateX: "0px", duration: 0.5 })
+    .addLabel("end");
+
+  gsap.set(".timeline-item", { opacity: 0, x: -20 });
+  ScrollTrigger.batch(".timeline-item", {
+    start: "-70px center",
+    end: "top center",
+    // markers: true,
+    onEnter: (batch) =>
+      gsap.to(batch, { opacity: 1, x: 0, autoAlpha: 1, stagger: 0.15 }),
+  });
+});
 </script>
